requirements.txt
aiofiles==23.2.1
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
arrow==1.2.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
crashtest==0.4.1
dataclasses-json==0.5.7
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
dnspython==2.4.2
dulwich==0.21.3
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.101.1
filelock==3.11.0
frozenlist==1.4.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.2.3
jedi==0.19.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
loguru==0.7.0
lxml==4.9.3
mangum==0.17.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.25.2
oauthlib==3.2.2
openai==1.12.0
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.2
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.39
protobuf==4.22.1
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.16.1
PyMuPDF==1.23.25
PyMuPDFb==1.23.25
pyodbc==4.0.35
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.3.0
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.1.1
rapidfuzz==2.15.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.5.2
tomli==2.0.1
tomlkit==0.11.7
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
trove-classifiers==2023.3.9
typing-inspect==0.8.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.21.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==12.0
xattr==0.10.1
yarl==1.9.2
zipp==3.15.0
